requirements.txt
tqdm==4.28.1
git+https://github.com/santi-pdp/ahoproc_tools.git
numpy==1.16.4
numba==0.48.0
scipy==1.1.0
librosa==0.6.3
SoundFile==0.10.2
torchaudio==0.4.0
torchvision==0.5.0
pysptk==0.1.16
matplotlib==3.0.2
python_speech_features==0.6
scikit_learn==0.20.3
tensorboardX==1.6
webrtcvad==2.0.10
cupy-cuda111
pynvrtc==8.0 
git+https://github.com/salesforce/pytorch-qrnn
git+https://github.com/detly/gammatone
git+https://github.com/pswietojanski/kaldi-io-for-python.git
